requirements.txt
# Core dependencies for AI customer support agent
google-generativeai==0.8.5
websockets==11.0.3
python-dotenv==1.0.0
pydantic==2.11.7
pydantic-settings==2.2.1
loguru==0.7.3
typing-extensions==4.14.1 